refactored processes
