--- epsilon_transformers/processes/rrxor.py
@@ -1,23 +1,0 @@
-import numpy as np
-
-from epsilon_transformers.processes.process import Process
-
-class RRXOR(Process):
-    def __init__(self, pR1=0.5, pR2=0.5):
-        self.pR1 = pR1
-        self.pR2 = pR2
-        super().__init__()
-
-    def _create_hmm(self):
-        T = np.zeros((2, 5, 5))
-        state_names = {'S': 0, '0': 1, '1': 2, 'T': 3, 'F': 4}
-        T[0, state_names['S'], state_names['0']] = self.pR1
-        T[1, state_names['S'], state_names['1']] = 1 - self.pR1
-        T[0, state_names['0'], state_names['F']] = self.pR2
-        T[1, state_names['0'], state_names['T']] = 1 - self.pR2
-        T[0, state_names['1'], state_names['T']] = self.pR2
-        T[1, state_names['1'], state_names['F']] = 1 - self.pR2
-        T[1, state_names['T'], state_names['S']] = 1.0
-        T[0, state_names['F'], state_names['S']] = 1.0
-
-        return T, state_names
