--- epsilon_transformers/processes/zero_one_random.py
@@ -1,19 +1,0 @@
-import numpy as np
-
-from epsilon_transformers.processes.process import Process
-
-
-class ZeroOneR(Process):
-    def __init__(self, prob_of_zero_from_r_state: float=0.5):
-        self.p = prob_of_zero_from_r_state
-        super().__init__()
-
-    def _create_hmm(self):
-        T = np.zeros((2, 3, 3))
-        state_names = {'0': 0, '1': 1, 'R': 2}
-        T[0, state_names['0'], state_names['1']] = 1.0
-        T[1, state_names['1'], state_names['R']] = 1.0
-        T[0, state_names['R'], state_names['0']] = self.p
-        T[1, state_names['R'], state_names['0']] = 1-self.p
-
-        return T, state_names
