+++ epsilon_transformers/train/run_sweeps.py
@@ -1,4 +1,4 @@
-from epsilon_transformers.processes.rrxor import RRXOR
+from epsilon_transformers.process.processes import RRXOR
 from epsilon_transformers.visualization import visualize_graph
 from epsilon_transformers.nn.simple_transformer import train_hooked_model
 from epsilon_transformers.markov_utilities import calculate_sequence_probabilities, compute_myopic_entropy_from_MSP, to_mixed_state_presentation
